chore: move utilitary functions out of ee folder (#31207)
