+++ apps/meteor/client/sidebar/footer/voip/hooks/useOmnichannelContactLabel.ts
@@ -2,7 +2,7 @@ import type { ICallerInfo } from '@rocket.chat/core-typings';
 import { useEndpoint } from '@rocket.chat/ui-contexts';
 import { useQuery } from '@tanstack/react-query';
 
-import { parseOutboundPhoneNumber } from '../../../../../ee/client/lib/voip/parseOutboundPhoneNumber';
+import { parseOutboundPhoneNumber } from '../../../../lib/voip/parseOutboundPhoneNumber';
 
 export const useOmnichannelContactLabel = (caller: ICallerInfo): string => {
 	const getContactBy = useEndpoint('GET', '/v1/omnichannel/contact.search');
